src/test/java/woowacourse/auth/application/AuthServiceTest.java
package woowacourse.auth.application;

import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import woowacourse.auth.dto.SignInDto;
import woowacourse.auth.dto.TokenResponseDto;
import woowacourse.shoppingcart.dto.request.SignUpDto;
import woowacourse.shoppingcart.service.CustomerService;

import static org.assertj.core.api.Assertions.assertThat;
import static woowacourse.fixture.Fixture.*;

@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
class AuthServiceTest {

    @Autowired
    private CustomerService customerService;

    @Autowired
    private AuthService authService;

    @Test
    @DisplayName("로그인 정보를 받아서 토큰을 반환한다.")
    void login() {
        //given
        final SignUpDto signUpDto = new SignUpDto(TEST_EMAIL, TEST_PASSWORD, TEST_USERNAME);
        customerService.signUp(signUpDto);
        final SignInDto signInDto = new SignInDto(TEST_EMAIL, TEST_PASSWORD);

        //when
        final TokenResponseDto token = authService.login(signInDto);

        //then
        final String subject = authService.extractEmail(token.getAccessToken());
        assertThat(subject).isEqualTo(TEST_EMAIL);
    }

}
